Add files via upload
This is your own novel and only you know how to write it. I'm doing it because I never want to be the person who looks back and wonders what could've happened if I wasn't too scary...and I have never want to be a popular person , I just been a kid who search a truth and looking for adventures.